Add files via upload
This is a script that allows the user to select a difficulty by first choosing the lowest and highest possible numbers the game will use. A random number is generated and the in a loop the user is prompted to make a guess, this guess is the compared to the actual number and a prompt is given depending on if they are correct, or higher/lower than the random number. During game play the number of guesses are counted and once the game is over then the game summary is written to a text file ('14_game_log.txt') with the current date & time.